# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MMJ Payroll System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /logos:
    get:
      operationId: listLogos
      tags: [logos]
      summary: List all uploaded logos
      responses:
        "200":
          description: List of logos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Logo"
    post:
      operationId: uploadLogo
      tags: [logos]
      summary: Upload a logo file (base64 encoded)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadLogoBody"
      responses:
        "201":
          description: Uploaded logo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Logo"
  /logos/{id}:
    get:
      operationId: getLogo
      tags: [logos]
      summary: Get a logo (returns base64 data)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Logo data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Logo"
    delete:
      operationId: deleteLogo
      tags: [logos]
      summary: Delete a logo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Invoices ─────────────────────────────────────────────────────────────────
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, all]
        - name: clientName
          in: query
          schema:
            type: string
        - name: periodId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice (manual or auto-generated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
  /invoices/generate-from-period:
    post:
      operationId: generateInvoiceFromPeriod
      tags: [invoices]
      summary: Auto-generate a draft invoice from pay period data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInvoiceFromPeriodBody"
      responses:
        "201":
          description: Generated invoice draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice (fields + line items)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/mark-sent:
    post:
      operationId: markInvoiceSent
      tags: [invoices]
      summary: Mark invoice as sent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/mark-paid:
    post:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/duplicate:
    post:
      operationId: duplicateInvoice
      tags: [invoices]
      summary: Duplicate an invoice as a new draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    BusinessProfile:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hstNumber:
          type: ["string", "null"]
        logoId:
          type: ["integer", "null"]
        logoUrl:
          type: ["string", "null"]
        isDefault:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, businessName, isDefault, createdAt, updatedAt]
...
    Logo:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        mimeType:
          type: string
        dataUrl:
          type: string
          description: Base64 data URL (data:image/png;base64,...)
        uploadedAt:
          type: string
      required: [id, filename, mimeType, dataUrl, uploadedAt]
    UploadLogoBody:
      type: object
      properties:
        filename:
          type: string
        mimeType:
          type: string
        dataUrl:
          type: string
          description: Base64 data URL
      required: [filename, mimeType, dataUrl]
    InvoiceLine:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        hours:
          type: ["number", "null"]
        rate:
          type: ["number", "null"]
        amount:
          type: number
        sortOrder:
          type: integer
      required: [id, description, amount, sortOrder]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        businessProfileId:
          type: ["integer", "null"]
        businessName:
          type: ["string", "null"]
        clientName:
          type: string
        clientAddress:
          type: ["string", "null"]
        periodId:
          type: ["integer", "null"]
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        serviceDescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        taxRate:
          type: ["number", "null"]
        subtotal:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [draft, sent, paid]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, invoiceNumber, clientName, invoiceDate, subtotal, taxAmount, total, status, createdAt, updatedAt]
    InvoiceDetail:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        businessProfileId:
          type: ["integer", "null"]
        businessName:
          type: ["string", "null"]
        clientName:
          type: string
        clientAddress:
          type: ["string", "null"]
        periodId:
          type: ["integer", "null"]
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        serviceDescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        taxRate:
          type: ["number", "null"]
        subtotal:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [draft, sent, paid]
        createdAt:
          type: string
        updatedAt:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLine"
        businessProfile:
          nullable: true
          $ref: "#/components/schemas/BusinessProfile"
      required: [id, invoiceNumber, clientName, invoiceDate, subtotal, taxAmount, total, status, createdAt, updatedAt, lineItems]
    CreateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: ["string", "null"]
        businessProfileId:
          type: ["integer", "null"]
        clientName:
          type: string
        clientAddress:
          type: ["string", "null"]
        periodId:
          type: ["integer", "null"]
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        serviceDescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        taxRate:
          type: ["number", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceLineBody"
      required: [clientName, invoiceDate]
    UpdateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: string
        businessProfileId:
          type: ["integer", "null"]
        clientName:
          type: string
        clientAddress:
          type: ["string", "null"]
        periodId:
          type: ["integer", "null"]
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        serviceDescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        taxRate:
          type: ["number", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceLineBody"
    CreateInvoiceLineBody:
      type: object
      properties:
        description:
          type: string
        hours:
          type: ["number", "null"]
        rate:
          type: ["number", "null"]
        amount:
          type: number
        sortOrder:
          type: integer
      required: [description, amount, sortOrder]
    GenerateInvoiceFromPeriodBody:
      type: object
      properties:
        periodId:
          type: integer
        hotelId:
          type: ["integer", "null"]
        businessProfileId:
          type: ["integer", "null"]
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        taxRate:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required: [periodId, invoiceDate]