# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MMJ Payroll System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInvoiceLineBody } from "./createInvoiceLineBody";

export interface CreateInvoiceBody {
  /** @nullable */
  invoiceNumber?: string | null;
  /** @nullable */
  businessProfileId?: number | null;
  clientName: string;
  /** @nullable */
  clientAddress?: string | null;
  /** @nullable */
  periodId?: number | null;
  invoiceDate: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  serviceDescription?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  taxRate?: number | null;
  lineItems?: CreateInvoiceLineBody[];
}