# lib/api-zod/src/generated/types/invoiceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MMJ Payroll System API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessProfile } from "./businessProfile";
import type { InvoiceDetailStatus } from "./invoiceDetailStatus";
import type { InvoiceLine } from "./invoiceLine";

export interface InvoiceDetail {
  id: number;
  invoiceNumber: string;
  /** @nullable */
  businessProfileId?: number | null;
  /** @nullable */
  businessName?: string | null;
  clientName: string;
  /** @nullable */
  clientAddress?: string | null;
  /** @nullable */
  periodId?: number | null;
  invoiceDate: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  serviceDescription?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  taxRate?: number | null;
  subtotal: number;
  taxAmount: number;
  total: number;
  status: InvoiceDetailStatus;
  createdAt: string;
  updatedAt: string;
  lineItems: InvoiceLine[];
  businessProfile?: BusinessProfile | null;
}